--- backend/src/main/java/com/morpheus/backend/BackendApplication.java
@@ -1,13 +1,0 @@
-package com.morpheus.backend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BackendApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BackendApplication.class, args);
-	}
-
-}
