--- backend/src/test/java/com/morpheus/backend/BackendApplicationTests.java
@@ -1,13 +1,0 @@
-package com.morpheus.backend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BackendApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
